start google api refactor by moving providers into subfolder - idea is to share more utils
