--- src/providers/googleShared.ts
@@ -1,43 +1,0 @@
-export const CHAT_MODELS = [
-  'aqa',
-  'chat-bison',
-  'chat-bison-32k',
-  'chat-bison-32k@001',
-  'chat-bison-32k@002',
-  'chat-bison@001',
-  'chat-bison@002',
-  'codechat-bison',
-  'codechat-bison-32k',
-  'codechat-bison-32k@001',
-  'codechat-bison-32k@002',
-  'codechat-bison@001',
-  'codechat-bison@002',
-  'gemini-1.0-pro',
-  'gemini-1.0-pro-001',
-  'gemini-1.0-pro-002',
-  'gemini-1.0-pro-vision',
-  'gemini-1.0-pro-vision-001',
-  'gemini-1.0-pro-vision-001',
-  'gemini-1.5-flash',
-  'gemini-1.5-flash-001',
-  'gemini-1.5-flash-002',
-  'gemini-1.5-flash-8b-exp-0924',
-  'gemini-1.5-flash-8b-exp-0827',
-  'gemini-1.5-flash-exp-0827',
-  'gemini-1.5-flash-preview-0514',
-  'gemini-1.5-flash-latest',
-  'gemini-1.5-flash-8b-latest',
-  'gemini-1.5-flash-8b',
-  'gemini-1.5-flash-8b-001',
-  'gemini-1.5-pro',
-  'gemini-1.5-pro-001',
-  'gemini-1.5-pro-002',
-  'gemini-1.5-pro-latest',
-  'gemini-1.5-pro-preview-0409',
-  'gemini-1.5-pro-preview-0514',
-  'gemini-1.5-pro-exp-0827',
-  'gemini-2.5-pro-exp-03-25',
-  'gemini-pro',
-  'gemini-pro-vision',
-  'gemini-ultra',
-];
