+++ src/providers/registry.ts
@@ -32,8 +32,9 @@ import { DatabricksMosaicAiChatCompletionProvider } from './databricks';
 import { EchoProvider } from './echo';
 import { FalImageGenerationProvider } from './fal';
 import { GolangProvider } from './golangCompletion';
-import { GoogleChatProvider } from './google';
-import { GoogleMMLiveProvider } from './googleMultimodalLive';
+import { GoogleChatProvider } from './google/google';
+import { GoogleMMLiveProvider } from './google/googleMultimodalLive';
+import { VertexChatProvider, VertexEmbeddingProvider } from './google/vertex';
 import { GroqProvider } from './groq';
 import { HttpProvider } from './http';
 import {
@@ -74,7 +75,6 @@ import { ScriptCompletionProvider } from './scriptCompletion';
 import { SequenceProvider } from './sequence';
 import { SimulatedUser } from './simulatedUser';
 import { createTogetherAiProvider } from './togetherai';
-import { VertexChatProvider, VertexEmbeddingProvider } from './vertex';
 import { VoyageEmbeddingProvider } from './voyage';
 import { WatsonXProvider } from './watsonx';
 import { WebhookProvider } from './webhook';
